+++ assets/js/viewobat.js
@@ -3,13 +3,7 @@ import { deleteConfirmation } from './action-buttons.js';
 import './dropdown.js';
 search('obat');
 
-document.getElementById('add-button').addEventListener('click', function () {
-	Swal.fire({
-		template: '#tambah-data-obat',
-	});
-});
-
-window.addEventListener('click', function (event) {
+window.addEventListener('click', (event) => {
 	if (event.target.classList.contains('detail')) {
 		let namaObat = event.target.dataset.namaobat;
 		let keterangan = event.target.dataset.keterangan;
@@ -42,16 +36,16 @@ window.addEventListener('click', function (event) {
 			showCloseButton: true,
 		});
 	}
-});
 
-document.querySelectorAll('.update').forEach(function (item) {
-	item.addEventListener('click', function (event) {});
-});
+	if (event.target.id == 'add-button') {
+		Swal.fire({
+			template: '#tambah-data-obat',
+		});
+	}
 
-document.querySelectorAll('#delete-btn').forEach(function (btn) {
-	btn.addEventListener('click', function () {
-		const name = this.dataset.name;
-		const value = this.dataset.idobat;
+	if (event.target.id == 'delete-btn') {
+		const name = event.target.dataset.name;
+		const value = event.target.dataset.idobat;
 		deleteConfirmation(name, value);
-	});
+	}
 });
