+++ assets/js/viewpelanggan.js
@@ -2,31 +2,25 @@ import search from './search.js';
 import { deleteConfirmation } from './action-buttons.js';
 import './dropdown.js';
 search('pelanggan');
-const recipePhotos = document.querySelectorAll('.foto-resep');
 
-recipePhotos.forEach(function (item) {
-	item.addEventListener('click', function (event) {
+window.addEventListener('click', (event) => {
+	if (event.target.id == 'add-button') {
 		Swal.fire({
-			imageUrl: '../assets/img/' + this.dataset.fotoresep,
+			template: '#tambah-data-pelanggan',
+		});
+	}
+
+	if (event.target.classList.contains('foto-resep')) {
+		Swal.fire({
+			imageUrl: '../assets/img/' + event.target.dataset.fotoresep,
 			padding: '1em',
 			showConfirmButton: false,
 		});
+	}
 
-		event.stopPropagation();
-		event.preventDefault();
-	});
-});
-
-document.querySelectorAll('#delete-btn').forEach(function (btn) {
-	btn.addEventListener('click', function () {
-		const name = this.dataset.name;
-		const value = this.dataset.idpelanggan;
+	if (event.target.id == 'delete-btn') {
+		const name = event.target.dataset.name;
+		const value = event.target.dataset.idpelanggan;
 		deleteConfirmation(name, value);
-	});
-});
-
-document.getElementById('add-button').addEventListener('click', function () {
-	Swal.fire({
-		template: '#tambah-data-pelanggan',
-	});
+	}
 });
